+++ package.json
@@ -16,7 +16,6 @@
         "@parcel/transformer-mdx": "^2.8.3",
         "@parcel/transformer-sass": "^2.8.3",
         "@parcel/transformer-webmanifest": "^2.8.3",
-        "@types/bootstrap": "^5.2.6",
         "@types/eslint": "^8.21.3",
         "@types/events": "^3.0.0",
         "@types/keyboardjs": "^2.5.1",
@@ -109,7 +108,6 @@
         "@sentry/react": "^7.45.0",
         "@sentry/rrweb": "^0.3.3",
         "@sentry/tracing": "^7.45.0",
-        "bootstrap": "^5.2.3",
         "browserslist": "^4.21.5",
         "comlink": "^4.4.1",
         "copy-image-clipboard": "^2.1.2",
@@ -127,7 +125,6 @@
         "prettier": "^2.8.7",
         "pwa-asset-generator": "^6.2.1",
         "react": "^18.2.0",
-        "react-bootstrap": "^2.7.2",
         "react-calendar": "^4.1.0",
         "react-dom": "^18.2.0",
         "react-icons": "^4.8.0",
